refactor: common 모듈과 global 모듈 통합
