+++ global/src/main/java/com/kernel/global/common/enumerate/PostStatus.java
@@ -1,4 +1,4 @@
-package com.kernel.common.global.enums;
+package com.kernel.global.common.enumerate;
 
 public enum PostStatus {
     TEMP("임시"),
